pyproject.toml
[tool.poetry]
name = "news-etl-pipeline"
version = "0.0.1"
description = "The goal of this project is to build a scalable and automated ETL pipeline that extracts news data from BBC News, stores it in a NoSQL database, and provides real-time monitoring of the scraping process to ensure reliability and performance."
authors = ["Rahul Shelke <rahul@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
pymongo = "^4.15.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-mock = "^3.15.1"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-v"
